apps/python-agents/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langchain==0.1.4
langchain-openai==0.0.5
langgraph==0.0.20
httpx==0.26.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
